moved mock lemonade api to its own file
