+++ apps/passport-server/src/services/genericIssuanceService.ts
@@ -6,7 +6,6 @@ import {
 import { SerializedPCD } from "@pcd/pcd-types";
 import { Router } from "express";
 import _ from "lodash";
-import { LemonadeDataMocker } from "../../test/lemonade/LemonadeDataMocker";
 import { logger } from "../util/logger";
 
 /**
@@ -80,60 +79,6 @@ export interface LemonadeUser {
   email: string;
   apiKey: string;
   name: string;
-}
-
-/**
- * TODO:
- * - multiple keys
- * - mutations
- */
-export class MockLemonadeAPI implements ILemonadeAPI {
-  private data: LemonadeDataMocker;
-
-  public constructor(data: LemonadeDataMocker) {
-    this.data = data;
-  }
-
-  public async loadEvents(apiKey: string): Promise<LemonadeEvent[]> {
-    const user = this.data.getUserByApiKey(apiKey);
-
-    if (!user) {
-      throw new Error(`no user with api key ${apiKey} found`);
-    }
-
-    return this.data.getUsersEvents(user.id);
-  }
-
-  public async checkinTicket(
-    apiKey: string,
-    eventId: string,
-    ticketId: string
-  ): Promise<void> {
-    const user = this.data.getUserByApiKey(apiKey);
-
-    if (!user) {
-      throw new Error(`no user with api key ${apiKey} found`);
-    }
-
-    const usersEvents = this.data.getUsersEvents(user.id);
-    const event = usersEvents.find((e) => e.id === eventId);
-
-    if (!event) {
-      throw new Error(`user ${user.id} cannot view event ${eventId}`);
-    }
-
-    const ticket = event.tickets.find((t) => t.id === ticketId);
-
-    if (!ticket) {
-      throw new Error(`could not find ticket with id ${ticketId}`);
-    }
-
-    if (ticket.checkedIn) {
-      throw new Error(`ticket ${ticket.id} is already checked in`);
-    }
-
-    ticket.checkedIn = true;
-  }
 }
 
 /**
